+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
  - Change terminology in `steadybit service init` command to comply with terminology found in the public documentation.
+ - Add a `steadybit service verify .steadybit.yml` command to compare the actual vs. desired resilience level.
 
 ## v0.4.0
  - Correct supported Node.js version to `>=16`.